Draws a Map with Earthquake Locations

eq_map plots the locations of earthquake events on a geographical map of one or multiple countries. Earthquake locations are plotted as red circles with labels and size scaled based on magnitude.

Usage

eq_map(data, annot_col, scaling_factor = 1)

Arguments

data

A data frame containing columns for "country", "magnitude", "longitude" and "latitude". The latter two must be values in [deg].

annot_col

A vector of values which will be used for labels.

scaling_factor

A numeric value which scales the point size. Defaults to 1.

Value

An interactive leaflet map.

Examples

library(dplyr)
# Plot earthquake events for a single country
southamerica %>%
  dplyr::filter(country == "CHILE" & lubridate::year(date) >= 1970) %>%
  eq_map(annot_col = "date", scaling_factor = 5)

# Plot earthquake events for multiple countries
southamerica %>%
  dplyr::filter(lubridate::year(date) > 2000) %>%
  eq_map(annot_col = "mag", scaling_factor = 5)